renamed clientRsvp.js to publicRsvp.js
